optimization: template form api unified
